docs: Update README
