+++ modules/financial_bot/tools/ui.py
@@ -1,14 +1,55 @@
 import logging
+from pathlib import Path
 from threading import Thread
 from typing import List
 
 import gradio as gr
 
-from financial_bot import load_bot
-
 logger = logging.getLogger(__name__)
 
+
+# === Load Bot ===
+
+def load_bot(
+    env_file_path: str = ".env",
+    logging_config_path: str = "logging.yaml",
+    model_cache_dir: str = "./model_cache",
+    embedding_model_device: str = "cuda:0",
+    debug: bool = False,
+):
+    """
+    Load the financial assistant bot in production or development mode based on the `debug` flag
+
+    production: the embedding model runs on GPU and the fine-tuned LLM is used.
+    dev: the embedding model runs on CPU and the fine-tuned LLM is mocked.
+    """
+
+    from financial_bot import initialize
+
+    # Be sure to initialize the environment variables before importing any other modules.
+    initialize(logging_config_path=logging_config_path, env_file_path=env_file_path)
+
+    from financial_bot import utils
+    from financial_bot.langchain_bot import FinancialBot
+
+    logger.info("#" * 100)
+    utils.log_available_gpu_memory()
+    utils.log_available_ram()
+    logger.info("#" * 100)
+
+    bot = FinancialBot(
+        model_cache_dir=Path(model_cache_dir) if model_cache_dir else None,
+        embedding_model_device=embedding_model_device,
+        debug=debug,
+    )
+
+    return bot
+
+
 bot = load_bot()
+
+
+# === Gradio Interface ===
 
 
 def predict(message: str, history: List[List[str]], about_me: str):
@@ -26,7 +67,12 @@ def predict(message: str, history: List[List[str]], about_me: str):
 
 demo = gr.ChatInterface(
     predict,
-    textbox=gr.Textbox(placeholder="Ask me a financial question", label="Financial question", container=False, scale=7),
+    textbox=gr.Textbox(
+        placeholder="Ask me a financial question",
+        label="Financial question",
+        container=False,
+        scale=7,
+    ),
     additional_inputs=[
         gr.Textbox(
             "I am a student and I have some money that I want to invest.",
